refactor(morpho-sdk): expose requirement helpers
