+++ packages/morpho-sdk/src/actions/requirements/encode/encodeErc20Permit2Transfer.ts
@@ -1,13 +1,13 @@
 import { getPermit2TransferFromTypedData } from "@morpho-org/blue-sdk-viem";
 import { deepFreeze, Time } from "@morpho-org/morpho-ts";
 import type { Address, WalletClient } from "viem";
+import { signAndVerifyTypedData } from "../../../helpers/signAndVerifyTypedData.js";
+import { validateRequirementSpender } from "../../../helpers/validateRequirementSpender.js";
 import type {
   Permit2TransferAction,
   Permit2TransferArgs,
   Requirement,
 } from "../../../types/index.js";
-import { signAndVerifyTypedData } from "../signAndVerifyTypedData.js";
-import { validateRequirementSpender } from "./validateRequirementSpender.js";
 
 /** Parameters for {@link encodeErc20Permit2Transfer}. */
 export interface EncodeErc20Permit2TransferParams {
